Fixes for gcc-3.0 warnings and errors. May have broken reintegration because of changed varargs handling.

+++ coda-src/venus/fso1.cc
@@ -817,7 +817,8 @@ void fsobj::ReplaceStatus(ViceStatus *vstat, vv_t *UpdateSet)
 	LOG(0, ("fsobj::ReplaceStatus: (%s), changed stat.length %d->%d\n",
 		FID_(&fid), stat.Length, vstat->Length));
 	if (IsFile())
-	    LocalSetAttr(-1, vstat->Length, -1, -1, -1);
+	    LocalSetAttr((unsigned long)-1, vstat->Length, (unsigned long)-1,
+			 (unsigned long)-1, (unsigned long)-1);
 	SetRcRights(RC_STATUS);
     }
